requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
pip==25.1.1
pysubs2==1.8.0
setuptools==58.0.4
